Refactor test generator script
