+++ .gitignore
@@ -34,6 +34,8 @@ playwright-report
 
 ./pages/api/script/*
 
+__pycache__
+
 scripts/exportData/exports
 
 jsonoutputs
