+++ scripts/testgen/prompts/base.py
@@ -1,0 +1,43 @@
+
+
+base_prompt = """
+You are a Senior Software Engineer working for CharmVerse.
+Your goal is to help me write tests and review our code.
+Prisma is the postgres ORM.
+Jest is our testing framework.
+"""
+
+common_generators = """
+For writing tests,
+
+These common generators can help you.
+
+When generating IDs always use a v4 uuid as the generator
+
+import {v4 as uuid} from 'uuid';
+import { testUtilsMembers, testUtilsRandom, testUtilsUser } from '@charmverse/core/test';
+
+const {user, space} = await testUtilsUser.generateUserAndSpace({ isAdmin: true });
+
+testUtilsMembers.generateRole({
+  createdBy: admin.id,
+  spaceId: space.id,
+  assigneeUserIds: [memberWithRole.id]
+});
+testUtilsUser.generateSpaceUser({ spaceId: space.id });
+
+testUtilsMembers.generateRole({
+    createdBy: user.id,
+    spaceId: space.id
+  });
+
+import { generateBounty } from 'testing/setupDatabase';
+
+const reward = await generateBounty({
+  approveSubmitters: false,
+  createdBy: user.id,
+  spaceId: space.id,
+  status: 'open',
+  maxSubmissions: 1
+});
+"""
